refactor: extract electron auto updater
